backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
requests>=2.26,<3.0
gunicorn>=20.1.0
django-cors-headers>=3.0
djangorestframework-authtoken>=2.0